Improve graph style and fix note link bugs
Now when a note is removed all note links this note was part in are also
removed from the database.

Also when updating the database now all notes are first inserted in the
database before inserting tags and links as the notes need to exist
already to be able to successfully insert all links (as the program
checks if both notes, which are part of the new link, exist, before
adding the link).